# swagger.yaml
openapi: 3.0.3
info:
  title: Music Platform API
  version: 1.0.0
  description: API for managing songs, albums, users, and messages with Google authentication
servers:
  - url: http://localhost:3001
paths:
...
  /api/users:
    get:
      summary: Get all users except the current user
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - No token provided or invalid token
  /api/users/messages/{userId}:
    get:
      summary: Get messages between current user and another user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID to fetch messages for
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Unauthorized - No token provided or invalid token
  /api/users/{userId}:
    get:
      summary: Get details of a specific user
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - No token provided or invalid token
        '404':
          description: User not found
    put:
      summary: Update user details
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  description: User's full name
                imageUrl:
                  type: string
                  description: URL of the user's profile image
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - No token provided or invalid token
        '403':
          description: Forbidden - Cannot edit another user's profile
        '404':
          description: User not found
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        fullName:
          type: string
        imageUrl:
          type: string
        googleId:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - fullName
        - imageUrl
        - googleId
        - email
    Message:
      type: object
      properties:
        _id:
          type: string
        senderId:
          type: string
        receiverId:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - senderId
        - receiverId
        - content
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT